refactor the code by separating resolvers
