+++ graphql/resolvers/booking.js
@@ -1,0 +1,48 @@
+const Booking = require("../../models/booking")
+const Event = require("../../models/event")
+const {transformBooking, transformEvent} = require("./shared")
+
+module.exports = {
+  bookings: async () => {
+    try {
+      const bookings = await Booking.find()
+      return bookings.map((booking) => {
+        return transformBooking(booking)
+        // return {
+        //   ...booking._doc,
+        //   _id: booking.id,
+        //   user: user.bind(this, booking._doc.user),
+        //   event: singleEvent.bind(this, booking._doc.event),
+        //   createdAt: dateToString(booking._doc.createdAt),
+        //   updatedAt: dateToString(booking._doc.createdAt),
+        // }
+      })
+    } catch (error) {
+      throw error
+    }
+  },
+
+  bookEvent: async (args) => {
+    try {
+      const fetchedEvent = await Event.findOne({_id: args.eventId})
+      const booking = new Booking({
+        user: "5f42e11d1a1a092bf85265ae",
+        event: fetchedEvent,
+      })
+      const result = await booking.save()
+      return transformBooking(result)
+    } catch (error) {
+      throw error
+    }
+  },
+  cancelBooking: async (args) => {
+    try {
+      const booking = await Booking.findById(args.bookingId).populate("event")
+      const event = transformEvent(booking.event)
+      await Booking.deleteOne({_id: args.bookingId})
+      return event
+    } catch (error) {
+      throw error
+    }
+  },
+}
