+++ helpers/date.js
@@ -1,0 +1,1 @@
+exports.dateToString = (date) => new Date(date).toISOString()
