+++ models/booking.js
@@ -2,13 +2,13 @@ const mongoose = require("mongoose")
 
 const Schema = mongoose.Schema
 
-const bookingSchema = new mongoose.Schema(
+const bookingSchema = new Schema(
   {
-    eventId: {
+    event: {
       type: Schema.Types.ObjectId,
       ref: "Event",
     },
-    userId: {
+    user: {
       type: Schema.Types.ObjectId,
       ref: "User",
     },
